tests/wpt/web-platform-tests/xhr/access-control-basic-allow-non-cors-safelisted-method.htm

<!DOCTYPE html>
<html>
  <head>
    <title>Tests cross-origin request with non-CORS-safelisted method</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/get-host-info.sub.js"></script>
  </head>
  <body>
    <script type="text/javascript">
    test(function() {
      const xhr = new XMLHttpRequest;

      xhr.open("PUT", get_host_info().HTTP_REMOTE_ORIGIN + "/xhr/resources/access-control-basic-put-allow.py", false);

      xhr.setRequestHeader("Content-Type", "text/plain; charset=UTF-8");

      xhr.send("PASS: PUT data received");

      assert_equals(xhr.responseText, "PASS: Cross-domain access allowed.\nPASS: PUT data received");
    }, "Allow PUT request");
    </script>
  </body>
</html>
